+++ gameserver/src/main/java/com/lhf/messages/in/CommandAdapter.java
@@ -6,15 +6,13 @@ import java.util.Optional;
 import java.util.StringJoiner;
 
 import com.lhf.messages.Command;
-import com.lhf.messages.CommandChainHandler;
-import com.lhf.messages.CommandContext;
 import com.lhf.messages.CommandMessage;
 import com.lhf.messages.grammar.Prepositions;
 
 public abstract class CommandAdapter {
     protected final Command command;
 
-    public CommandAdapter(Command command) {
+    protected CommandAdapter(Command command) {
         this.command = command;
     }
 
@@ -68,74 +66,6 @@ public abstract class CommandAdapter {
         sj.add("Valid:").add(this.isValid().toString());
         sj.add("Payload:").add(this.getWhole());
         return this.command.toString();
-    }
-
-    public static CommandContext.Reply dispatch(CommandContext ctx, Command command, CommandChainHandler handler) {
-        if (ctx == null || command == null || handler == null) {
-            throw new IllegalArgumentException(
-                    String.format("Cannot use a null context, command, or handler: %s, %s, %s", ctx, command, handler));
-        }
-        CommandMessage commandType = command.getType();
-        if (commandType == null) {
-            throw new IllegalStateException(String.format("Command '%s' must not have a null type", command));
-        }
-        switch (commandType) {
-            case ATTACK:
-                return handler.handleInCommand(ctx, new AttackMessage(command));
-            case CAST:
-                return handler.handleInCommand(ctx, new CastMessage(command));
-            case CREATE:
-                return handler.handleInCommand(ctx, new CreateInMessage(command));
-            case DROP:
-                return handler.handleInCommand(ctx, new DropMessage(command));
-            case EQUIP:
-                return handler.handleInCommand(ctx, new EquipMessage(command));
-            case EXIT:
-                return handler.handleInCommand(ctx, new ExitMessage(command));
-            case FOLLOW:
-                return handler.handleInCommand(ctx, new FollowMessage(command));
-            case GO:
-                return handler.handleInCommand(ctx, new GoMessage(command));
-            case HELP:
-                return handler.handleInCommand(ctx, new HelpInMessage(command));
-            case INTERACT:
-                return handler.handleInCommand(ctx, new InteractMessage(command));
-            case INVENTORY:
-                return handler.handleInCommand(ctx, new InventoryMessage(command));
-            case LEWD:
-                return handler.handleInCommand(ctx, new LewdInMessage(command));
-            case PASS:
-                return handler.handleInCommand(ctx, new PassMessage(command));
-            case PLAYERS:
-                return handler.handleInCommand(ctx, new ListPlayersMessage(command));
-            case REPEAT:
-                return handler.handleInCommand(ctx, new RepeatInMessage(command));
-            case REST:
-                return handler.handleInCommand(ctx, new RestMessage(command));
-            case SAY:
-                return handler.handleInCommand(ctx, new SayMessage(command));
-            case SEE:
-                return handler.handleInCommand(ctx, new SeeMessage(command));
-            case SHOUT:
-                return handler.handleInCommand(ctx, new ShoutMessage(command));
-            case SPELLBOOK:
-                return handler.handleInCommand(ctx, new SpellbookMessage(command));
-            case STATS:
-                return handler.handleInCommand(ctx, new StatsInMessage(command));
-            case STATUS:
-                return handler.handleInCommand(ctx, new StatusMessage(command));
-            case TAKE:
-                return handler.handleInCommand(ctx, new TakeMessage(command));
-            case UNEQUIP:
-                return handler.handleInCommand(ctx, new UnequipMessage(command));
-            case USE:
-                return handler.handleInCommand(ctx, new UseMessage(command));
-            default:
-                throw new UnsupportedOperationException(
-                        String.format("Need adapter for command type %s", commandType.toString()));
-
-        }
-
     }
 
     public static Optional<Command> fromString(String payload) {
